requirements.txt
annotated-types==0.7.0
anyio==4.9.0
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
mongoengine==0.29.1
numpy==2.3.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.3.0
pdfminer.six==20250506
pdftotext==3.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==4.13.0
pypdf==5.6.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
